Add resume version and logo
There are a lot of copies of my resume out in the wild. Adding a version
number to the bottom of the page immediately invalidates any previous
versions.

I also like the 'ok' emoji, so I added that as a nice conclusion mark.
This won't show up in PDF copies because I'm not fucking with LaTeX.